pyproject.toml
[tool.poetry]
name = "owid-grapher-py"
version = "0.1.0"
description = "OWID charts for rendering in Jupyter notebooks."
authors = ["Lars Yencken <lars@yencken.org>"]
packages = [{ include = "owid" }]

[tool.poetry.dependencies]
python = "^3.9"
dataclasses-json = "^0.5.4"
pandas = "^1.3.0"
requests = "^2.25.1"
python-dateutil = "^2.8.1"

[tool.poetry.dev-dependencies]
watchdog = "^2.1.3"
PyYAML = "^5.4.1"
argh = "^0.26.2"
pytest = "^6.2.4"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pyright]
include = ["owid/**", "tests/**"]
